move lambda runtime trait to api crate
