+++ src/api/mod.rs
@@ -4,6 +4,8 @@ mod env;
 mod handler;
 /// Defines the interface used to read a response from the Lambda API.
 mod response;
+/// Defines the lambda runtime interface.
+mod runtime;
 /// Defines the [`crate::transport::Transport`] abstraction used to support multiple HTTP backends.
 mod transport;
 
@@ -11,4 +13,5 @@ pub use crate::api::context::*;
 pub use crate::api::env::*;
 pub use crate::api::handler::*;
 pub use crate::api::response::*;
+pub use crate::api::runtime::*;
 pub use crate::api::transport::*;
